+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- A new command `lak list accounts` that allows printing account values and
+percentages by accounts or by account types.
 ### Fixed
 - The spinner chars were not showing properly on MS Windows 11. Changed to
 a simpler spinner.